Added Lots
Sorry I forgot to hit submit last night, I have been working a lot with this program and the base of it is all finished as of now. Right now I am going through it and playing it to see where I can make improvements and where I can see any bugs that pop up.